Moved check for method not implemented or supported to extension method of OperationBlockAnalysisContext
